# perf(http): cache webidl.converters lookups in ext/fetch/23_response.js