Code:

```js
Deno.bench("noop", () => {
  return 1;
})

Deno.bench("response", () => {
  const responses = [];
  for (var i = 0; i < 1024 * 1024; i++) {
    responses.push(new Response(i.toString()));
  }
  return responses;
});
```

Before:

```
response           572.2 ms           1.7 (443.7 ms … 752.5 ms) 679.6 ms 752.5 ms 752.5 ms
```

After:

```
response           543.3 ms           1.8 (369.5 ms … 721.8 ms) 606.3 ms 721.8 ms 721.8 ms
```